Included createUnauthorizedAIFileException method and refactored code
