+++ Standard.AI.OpenAI/Services/Foundations/AIFiles/AIFileService.Exceptions.cs
@@ -45,8 +45,7 @@ namespace Standard.AI.OpenAI.Services.Foundations.AIFiles
             catch (HttpResponseUnauthorizedException httpResponseUnauthorizedException)
             {
                 var unauthorizedAIFileException =
-                    new UnauthorizedAIFileException(
-                        message: "Unauthorized AI file request, fix errors and try again.", 
+                    createUnauthorizedAIFileException(
                         httpResponseUnauthorizedException);
 
                 throw createAIFileDependencyException(
@@ -55,8 +54,7 @@ namespace Standard.AI.OpenAI.Services.Foundations.AIFiles
             catch (HttpResponseForbiddenException httpResponseForbiddenException)
             {
                 var unauthorizedAIFileException =
-                    new UnauthorizedAIFileException(
-                        message: "Unauthorized AI file request, fix errors and try again.", 
+                    createUnauthorizedAIFileException(
                         httpResponseForbiddenException);
 
                 throw createAIFileDependencyException(
@@ -129,8 +127,7 @@ namespace Standard.AI.OpenAI.Services.Foundations.AIFiles
             catch (HttpResponseUnauthorizedException httpResponseUnauthorizedException)
             {
                 var unauthorizedAIFileException =
-                    new UnauthorizedAIFileException(
-                        message: "Unauthorized AI file request, fix errors and try again.", 
+                    createUnauthorizedAIFileException(
                         httpResponseUnauthorizedException);
 
                 throw createAIFileDependencyException(
@@ -139,8 +136,7 @@ namespace Standard.AI.OpenAI.Services.Foundations.AIFiles
             catch (HttpResponseForbiddenException httpResponseForbiddenException)
             {
                 var unauthorizedAIFileException =
-                    new UnauthorizedAIFileException(
-                        message: "Unauthorized AI file request, fix errors and try again.", 
+                    createUnauthorizedAIFileException(
                         httpResponseForbiddenException);
 
                 throw createAIFileDependencyException(
@@ -173,62 +169,6 @@ namespace Standard.AI.OpenAI.Services.Foundations.AIFiles
                 throw createAIFileServiceException(
                    failedAIFileServiceException);
             }
-        }
-
-        private static AIFileDependencyException createAIFileDependencyException(Xeption innerException)
-        {
-            return new AIFileDependencyException(
-                message: "AI file dependency error occurred, contact support.",
-                innerException);
-        }
-
-        private static AIFileDependencyValidationException createAIFileDependencyValidationException(Xeption innerException)
-        {
-            return new AIFileDependencyValidationException(
-                message: "AI file dependency validation error occurred, contact support.",
-                innerException);
-        }
-
-        private static AIFileServiceException createAIFileServiceException(Xeption innerException)
-        {
-            return new AIFileServiceException(
-                message: "AI file service error occurred, contact support.",
-                innerException);
-        }
-
-        private static AIFileValidationException createAIFileValidationException(Xeption innerException)
-        {
-            throw new AIFileValidationException(
-                message: "AI file validation error occurred, fix errors and try again.",
-                innerException);
-        }
-
-        private static ExcessiveCallAIFileException createExcessiveCallAIFileException(Xeption innerException)
-        {
-            return new ExcessiveCallAIFileException(
-                message: "Excessive call error occurred, limit your calls.",
-                innerException);
-        }
-
-        private static FailedAIFileServiceException createFailedAIFileServiceException(Exception innerException)
-        {
-            return new FailedAIFileServiceException(
-                message: "Failed AI file service error occurred, contact support.",
-                innerException);
-        }
-
-        private static FailedServerAIFileException createFailedServerAIFileException(Exception innerException)
-        {
-            return new FailedServerAIFileException(
-                message: "Failed AI file server error occurred, contact support.",
-                innerException);
-        }
-
-        private static InvalidConfigurationAIFileException createInvalidConfigurationAIFileException(Exception innerException)
-        {
-            return new InvalidConfigurationAIFileException(
-                message: "Invalid AI file configuration error occurred, contact support.",
-                innerException);
         }
 
     }
